Applied some of the SGI fixes. Don't know about the     toolbar one.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@5626 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/unix/dialup.cpp
@@ -566,7 +566,7 @@ wxDialUpManagerImpl::CheckIfconfig(void)
                     || strstr(output,"sl")  // slip
                     || strstr(output,"pl"); // plip
 #elif defined(__SGI__)  // IRIX
-               rc = strstr(output, "ppp"); // PPP
+               rc = (int) strstr(output, "ppp"); // PPP
 #elif defined(__HPUX__)
                // if could run ifconfig on interface, then it exists
                rc = TRUE;
